requirements.txt
# Core dependencies
codecarbon>=2.3.0
web3>=6.0.0
psutil>=5.9.0
GPUtil>=1.4.0

# API and database
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
